+++ packages/zkdb/src/sdk/global/index.ts
@@ -1,0 +1,1 @@
+export { ZKDatabaseClient } from './zkdatabase-client.js'
